Functions for connection, basic commands, constants and production
The entire functionality for how to connect to the machine added in the file connection. This includes the ability to start a new session and end a session. 

Controller act as a single entry point for the API controller, but that function isnt added yet. It will controll all the functions of the nodeopcua. 

Production has at the moment only the basic commands for starting, stopping and resetting the machine with a production. 

Commands has the 3 basic commands that is used a lot. Getting the current state of the machine, changeing the state and confirming the change.