Helper function jsonBuilder, MachineNotReadyError
New helper function helps builds json objects to return to the caller. This cuts down on the duplicates in the code. 

A new class named MachineNotReadyError extends the error class in javascript and is a cutom error type. It will allow the startProduction to throw an error when it isnt in state 4 and thus return a message.