Refactor the catch and finally in production.js
Finally has been made to a much more readable one line if statement instead of the 3 lines before. 

Catch statement has lost all its if statement for the different kinds of custom error codes and now just use the super power of having our own custom super error class. Using a tenary operator it checks if it is an instance of CustomError(Super bad name), if it is it uses the super power of err.toJSON() which returns a json object. If it isnt some unkown error has defeated the machine and it returns a json object using bobTheBuilder()