Now Shows Simple Batch Details
The /details/id link now takes the URL, splits it into an array of strings, takes the last/[4] element and uses that as the ID to enter into the API call. Then it receives the data that the database sends back up, and writes it to the data React.state. Then, each row in the left table writes the corresponding data to the cell. 

Next step: 
- show the values of temperature, humidity and vibrations on the graph
- set up the window.print() functionality on the export button