Simulation page (pretty much) done!
The simulation now writes the calculated value to the bottom cell if the returned value is a number. If the returned value is instead the bad request response, it will instead write "Error, invalid input" to the cell. 

The Reset button is the only thing that can still be corrected here. It currently resets the values within the React.state, however it does not clear the text fields. This can be corrected, but low priority as of now.